[BE] test: Member 패키지 presentation을 controller/integration 테스트로 분리 #598
